+++ frontend/src/pages/Chatting/ChattingRoomPage/components/ChatBubble/ChatBubble.style.ts
