+++ frontend/src/pages/Chatting/ChattingRoomPage/components/ChatBubble/ChatBubble.tsx
