+++ frontend/src/pages/Chatting/ChattingRoomPage/components/ChattingFooter/ChattingFooter.style.ts
